added caching mechanism to FindBin

+++ internal/pkg/util/bin/bin.go
@@ -11,6 +11,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"sync"
 
 	"github.com/pkg/errors"
 	"github.com/sylabs/singularity/internal/pkg/buildcfg"
@@ -19,49 +20,67 @@ import (
 	"github.com/sylabs/singularity/pkg/util/singularityconf"
 )
 
+var findBinCache sync.Map
+
 // FindBin returns the path to the named binary, or an error if it is not found.
 func FindBin(name string) (path string, err error) {
+	pathEntry, ok := findBinCache.Load(name)
+	if ok {
+		path = pathEntry.(string)
+		return path, nil
+	}
+
 	switch name {
 	// Basic system executables that we assume are always on PATH
 	case "true", "mkfs.ext3", "cp", "rm", "dd", "truncate":
-		return findOnPath(name)
+		path, err = findOnPath(name)
 	// Bootstrap related executables that we assume are on PATH
 	case "mount", "mknod", "debootstrap", "pacstrap", "dnf", "yum", "rpm", "curl", "uname", "zypper", "SUSEConnect", "rpmkeys", "proot":
-		return findOnPath(name)
+		path, err = findOnPath(name)
 	// Configurable executables that are found at build time, can be overridden
 	// in singularity.conf. If config value is "" will look on PATH.
 	case "unsquashfs", "mksquashfs", "go":
-		return findFromConfigOrPath(name)
+		path, err = findFromConfigOrPath(name)
 	// distro provided setUID executables that are used in the fakeroot flow to setup subuid/subgid mappings
 	case "newuidmap", "newgidmap":
-		return findOnPath(name)
+		path, err = findOnPath(name)
 	// distro provided OCI runtime
 	case "crun", "runc":
-		return findOnPath(name)
+		path, err = findOnPath(name)
 	// our, or distro provided conmon
 	case "conmon":
 		if buildcfg.CONMON_LIBEXEC == 1 {
-			return filepath.Join(buildcfg.LIBEXECDIR, "singularity", "bin", name), nil
+			path, err = filepath.Join(buildcfg.LIBEXECDIR, "singularity", "bin", name), nil
+		} else {
+			path, err = findOnPath(name)
 		}
-		return findOnPath(name)
 	// cryptsetup & nvidia-container-cli paths must be explicitly specified
 	// They are called as root from the RPC server in a setuid install, so this
 	// limits to sysadmin controlled paths.
 	// ldconfig is invoked by nvidia-container-cli, so must be trusted also.
 	case "cryptsetup", "ldconfig", "nvidia-container-cli":
-		return findFromConfigOnly(name)
+		path, err = findFromConfigOnly(name)
 		// distro provided fusermount for unpriv SIF mount
 	case "fusermount":
-		return findOnPath(name)
+		path, err = findOnPath(name)
 	// distro provided squashfuse for unpriv SIF mount and OCI-mode bare-image
 	// overlay
 	case "squashfuse":
-		return findOnPath(name)
+		path, err = findOnPath(name)
 	// fuse2fs for OCI-mode bare-image overlay
 	case "fuse2fs":
-		return findOnPath(name)
+		path, err = findOnPath(name)
+	default:
+		err = fmt.Errorf("unknown executable name %q", name)
 	}
-	return "", fmt.Errorf("unknown executable name %q", name)
+
+	if err != nil {
+		return "", err
+	}
+
+	findBinCache.Store(name, path)
+
+	return path, nil
 }
 
 // findOnPath performs a simple search on PATH for the named executable, returning its full path.
